more robust iccup crawler

+++ iccup/crawl.py
@@ -1,6 +1,4 @@
-import urllib2, re, time
-
-# ugly code to refactor
+import urllib2, re, time, os
 
 fullr = r'\<div>[PZT] v [PZT]\</div>.+?replays/download[^"]+'
 rr = r'replays/download[^"]+'
@@ -23,17 +21,19 @@ def dlreps(p):
             replay = re.findall(rr, fullreplay)[0]
             mu = find_mu(fullreplay)
             repfile = 'replays/' + mu + '/' + re.search(r'(\d+)\.html', replay).group(1) + '.rep'
-            output = open(repfile, 'w')
-            output.write(urllib2.urlopen(
-                urllib2.urlopen(prefix + '/' + replay).geturl()).read())
-            output.close()
-            print 'saved', repfile
+            if not os.path.exists(repfile):
+                output = open(repfile, 'w')
+                output.write(urllib2.urlopen(
+                    urllib2.urlopen(prefix + '/' + replay).geturl()).read())
+                output.close()
+                print 'saved', repfile
         except:
             print 'failed saving', replay
-            failed.append(fullreplay) 
+            failed.append((fullreplay, 0, len(failed)+1)) 
 
 if __name__ == '__main__':
-    page = urllib2.urlopen(prefix + "/starcraft/replays.html").read()
+    #page = urllib2.urlopen(prefix + "/starcraft/replays.html").read()
+    page = urllib2.urlopen(prefix + "/starcraft/replays/user.html").read()
     dlreps(page)
     mnext = re.search(rnext, page)
     print mnext
@@ -46,11 +46,29 @@ if __name__ == '__main__':
             print mnext
             print mnext.group(1)
         except urllib2.HTTPError:
-            time.sleep(10)
-        for fail in failed:
+            time.sleep(5)
+        for (i, fail) in enumerate(failed): 
+            if fail[1] < (fail[2] + 1):
+                failed[i] = (failed[i][0], failed[i][1]+1, failed[i][2])
+                try:
+                    replay = re.findall(rr, fail[0])[0]
+                    mu = find_mu(fail[0])
+                    repfile = 'replays/' + mu + '/' + re.search(r'(\d+)\.html', replay).group(1) + '.rep'
+                    output = open(repfile, 'w')
+                    output.write(urllib2.urlopen(
+                        urllib2.urlopen(prefix + '/' + replay).geturl()).read())
+                    output.close()
+                    print 'saved', repfile
+                except:
+                    print 'failed again with', replay
+            else:
+                del failed[i]
+    for (i, fail) in enumerate(failed):
+        if fail[1] < (fail[2] + 1):
+            failed[i] = (failed[i][0], failed[i][1]+1, failed[i][2])
             try:
-                replay = re.findall(rr, fail)[0]
-                mu = find_mu(fail)
+                replay = re.findall(rr, fail[0])[0]
+                mu = find_mu(fail[0])
                 repfile = 'replays/' + mu + '/' + re.search(r'(\d+)\.html', replay).group(1) + '.rep'
                 output = open(repfile, 'w')
                 output.write(urllib2.urlopen(
@@ -59,20 +77,9 @@ if __name__ == '__main__':
                 print 'saved', repfile
             except:
                 print 'failed again with', replay
-                time.sleep(10)
-    for fail in failed:
-        try:
-            replay = re.findall(rr, fail)[0]
-            mu = find_mu(fail)
-            repfile = 'replays/' + mu + '/' + re.search(r'(\d+)\.html', replay).group(1) + '.rep'
-            output = open(repfile, 'w')
-            output.write(urllib2.urlopen(
-                urllib2.urlopen(prefix + '/' + replay).geturl()).read())
-            output.close()
-            print 'saved', repfile
-        except:
-            print 'failed again with', replay
-            time.sleep(10)
+                time.sleep(2)
+        else:
+            del failed[i]
     print 'finally failed with:', len(failed)
